CA MARCHE

+++ EasySaveConsole/Model/Logs.cs
@@ -12,7 +12,7 @@ namespace EasySaveConsole.Model
         public String TaskName { get; set; } //nom de la tache
         public String SourceFileAddress { get; set; } //adresse fichier source
         public String DestinationFileAddress { get; set; } //adress fichier de destination
-        public byte FileSize { get; set; } //taille fichier
+        public long FileSize { get; set; } //taille fichier
         public TimeSpan TransferTime { get; set; } //temps de transfert
     }
 }
